Correctly emit PaymentReceivedEvent on rebalances
Previously we did not do this under bad assumptions. We were able to do
this only because in tests we could guaranatee the payment ids were the
same across trusted and self custodial because we were using ldk-node's
payment id. However, with actual trusted backends we can't do this so we
have no way to see if the payment is a rebalance until after we've
received it.

To prevent this from happening again we now mangle the id in
DummyTrustedWallet so we don't accidentially have tests assuming that.