Add log facade option
Previously, we only had the option of a log file and we'd always drop
the trusted wallet's logs on the floor because they do not implement
LDK's Logger interface. We add the option to have logs output using the
`log` crate instead of to a file. This allows the user to handle the
logs in the way they see fit and can integrate better with other rust
projects.