Fix race condition in rebalances
Before we would return a successful rebalance on when we received the
final payment, but we did not wait for the trusted wallet to get the
succcess back. This could cause a race where if we called
list_transactions() after the rebalance but before the trusted wallet
got the success event, list_transactions would not correctly reflect the
rebalance and would hit a debug_assert.

This fixes by waiting for the payment success event on the trusted side.