Provide at least marginally useful `TrustedError`s in bindings
Given there is likely context in `TrustedError`s its helpful to
have at least a `String` for them. Would be nice for them to be
fully structured with helpful error messages, but this is easy and
at least is better than nothing (and we're somewhat limited by
upstream libraries anyway).