Handle internal transfers completed after shutdown
Currently we only process setting `TxMetadata` for internal
transfers if they complete in-line while the app is open and
running the rebalance logic. If, however, an internal transfer is
initiated but then does not complete the metadata is lost and
instead our transaction list includes a LN payment received and no
information about the trusted transaction at all.

Instead, here, we track enough information in
`TxType::PendingRebalance` to match the trusted transfer with an
LN transaction and then do so in `list_transactions`, updating
metadata as appropriate if we find a match.

We handle upgrades from previous versions of Orange gracefully,
though I'm not sure if we really need to do that yet.