Use our own tokio runtime rather than the UniFFI-default one
When we use `#[uniffi::export(async_runtime = "tokio")]` under the
hood it uses the `async-compat` crate to build a single-threaded
tokio runtime and then spawn a single thread to poll background
tasks.

Because `ldk-node` requires a multi-threaded runtime to spawn
blocking tasks, however, we build a fresh tokio runtime in the
`Wallet::new` constructor and let it run in the background as well.
This causes thread blowup and is a bit confusing as we end up
polling futures on multiple tokio runtimes, some of which are
kinda-single-threaded.

Instead, we just copy the 20 relevant lines out of `async-compat`
and spawn the `Wallet` futures on the tokio runtime we built for
`ldk-node`.